Split draw functions in their own file
